move test contexts to base.py
